Adding color space option
NEW : a combobox allows you to choose how to interpolate colors between
points on colormaps. When changed, it calls the colormap change
function, that will read the combobox and set the wanted colorspace.
BUG : there is still a bug when you select a mesh, change the
colorspace, then select an other mesh : the combobox does not update, I
did not find a way to read the colorspace like I did with the colormap
yet. Maybe I'll have to save it for each surface.