COMP: Simplify configuration of CLI output and install directories
By not explicitly passing *_OUTPUT_DIRECTORY and INSTALL_*_DESTINATION
parameters to SEMMacroBuildCLI, the default configured into the SlicerExecutionModel
are used.

It turns out that when the project is build as "stand-alone", the default
value are the bin and lib.

And when building as a Slicer extension, these value are set to match Slicer
convention (i.e. lib/Slicer-X.Y/cli-modules).

Co-authored-by: Jean-Christophe Fillion-Robin <JChris.FillionR@kitware.com>