COMP: Fix usage as a Slicer extension from a build directory
The launcher was built under the expectation it was used from an
installed tree. It was looking for ShapePopulationViewer executable
in a nonexistent directory named "hidden-cli-modules".

Since this project is not depending on any external projects when
built as a Slicer extension, there is no need to use a launcher.

This commit enables the use of the CLI module from a build tree by
removing the launcher.

Co-authored-by: Jean-Christophe Fillion-Robin <JChris.FillionR@kitware.com>