COMP: Simplify build system removing obsolete code related to VTK 5 or 6
This code is not needed anymore since VTK7 is required for the
stand-alone build, and also for the Slicer extension build.

Remove files associated with obsolete patch step that was used
with VTK 5 or 6.

Remove VTK_built_step file that became obsolete in 63fc395 (BUG: Updated
the options to build VTK and the list of the modules that are required by
ShapePopulationViewer)

Co-authored-by: Jean-Christophe Fillion-Robin <JChris.FillionR@kitware.com>